% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distns.R
\name{evaluateBetaDistribution}
\alias{evaluateBetaDistribution}
\title{Evaluate a fitted beta distribution}
\usage{
evaluateBetaDistribution(x, params, log = FALSE)
}
\arguments{
\item{x}{A numeric vector of values, in the range \eqn{[0,1]}, at which to
evaluate the density.}

\item{params}{A \code{betaDistribution} object.}

\item{log}{Boolean value: if \code{TRUE}, log-densities are returned.}
}
\value{
A numeric vector of (log-)densities.
}
\description{
This function evaluates the density of a beta distribution previously fit
with \code{\link[=fitBetaDistribution]{fitBetaDistribution()}} or \code{\link[=fitRegularisedBetaDistribution]{fitRegularisedBetaDistribution()}}.
}
\author{
Jon Clayden
}
