% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distns.R
\name{evaluateDistribution}
\alias{evaluateDistribution}
\title{Evaluate a fitted distribution}
\usage{
evaluateDistribution(x, params, log = FALSE)
}
\arguments{
\item{x}{A numeric vector of values at which to evaluate the density.}

\item{params}{A \code{betaDistribution}, \code{gaussianDistribution} or
\code{multinomialDistribution} object.}

\item{log}{Boolean value: if \code{TRUE}, log-densities are returned.}
}
\value{
A numeric vector of (log-)densities.
}
\description{
This function evaluates the density of a distribution object previously
fitted by one of \code{\link[=fitBetaDistribution]{fitBetaDistribution()}}, \code{\link[=fitGaussianDistribution]{fitGaussianDistribution()}} or
\code{\link[=fitMultinomialDistribution]{fitMultinomialDistribution()}} (or their variants), dispatching to the
appropriate specific evaluation function based on the class of \code{params}.
}
\author{
Jon Clayden
}
