% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distns.R
\name{evaluateGaussianDistribution}
\alias{evaluateGaussianDistribution}
\title{Evaluate a fitted Gaussian distribution}
\usage{
evaluateGaussianDistribution(x, params, log = FALSE)
}
\arguments{
\item{x}{A numeric vector of values at which to evaluate the density.}

\item{params}{A \code{gaussianDistribution} object.}

\item{log}{Boolean value: if \code{TRUE}, log-densities are returned.}
}
\value{
A numeric vector of (log-)densities.
}
\description{
This function evaluates the density of a Gaussian distribution previously
fit with \code{\link[=fitGaussianDistribution]{fitGaussianDistribution()}}.
}
\author{
Jon Clayden
}
