% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distns.R
\name{evaluateMultinomialDistribution}
\alias{evaluateMultinomialDistribution}
\title{Evaluate a fitted multinomial distribution}
\usage{
evaluateMultinomialDistribution(x, params, log = FALSE)
}
\arguments{
\item{x}{A single number, one of the values in \code{params$values} (if not,
the nearest value is used, with a warning), or a numeric vector of
frequencies of the same length as \code{params$probs}. \code{NA} is returned
unmodified.}

\item{params}{A \code{multinomialDistribution} object.}

\item{log}{Boolean value: if \code{TRUE}, the log-probability is returned.}
}
\value{
A single (log-)probability.
}
\description{
This function evaluates the probability mass of a multinomial
distribution, previously fit with \code{\link[=fitMultinomialDistribution]{fitMultinomialDistribution()}}, at a
single value or a full vector of observed frequencies.
}
\author{
Jon Clayden
}
