% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distns.R
\name{fitGaussianDistribution}
\alias{fitGaussianDistribution}
\title{Fit a Gaussian distribution}
\usage{
fitGaussianDistribution(data, mu = NULL, sigma = NULL)
}
\arguments{
\item{data}{A numeric vector of data values.}

\item{mu}{The mean of the distribution, or \code{NULL} to estimate it from
\code{data}.}

\item{sigma}{The standard deviation of the distribution, or \code{NULL} to
estimate it from \code{data}.}
}
\value{
A \code{gaussianDistribution} object, a list with elements \code{mu} and
\code{sigma}.
}
\description{
This function fits a Gaussian (normal) distribution to a vector of data by
maximum likelihood, i.e. using the sample mean and (biased) sample
standard deviation.
}
\author{
Jon Clayden
}
