% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getNTResource}
\alias{getNTResource}
\title{Read a PNT resource from file}
\usage{
getNTResource(type, options = NULL)
}
\arguments{
\item{type}{A string, one of \code{"reference"}, \code{"model"} or \code{"results"},
giving the type of resource required.}

\item{options}{A named list of elements used to construct the resource's
file name; see \code{\link[=getFileNameForNTResource]{getFileNameForNTResource()}}.}
}
\value{
The deserialised resource object: a \link{ReferenceTract} for
\code{type="reference"}, an \link{UninformativeTractModel} or \link{MatchingTractModel}
for \code{type="model"}, or a \link{ProbabilisticNTResults} object for
\code{type="results"}. The result is returned invisibly.
}
\description{
This function locates, using \code{\link[=getFileNameForNTResource]{getFileNameForNTResource()}}, and deserialises
a probabilistic neighbourhood tractography (PNT) reference tract, matching
model or results object.
}
\author{
Jon Clayden
}
