% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_bspline_tract.R
\name{getPointsForTract}
\alias{getPointsForTract}
\title{Extract representative points from a B-spline tract}
\usage{
getPointsForTract(tract, pointType = c("control", "knot"))
}
\arguments{
\item{tract}{A \link{BSplineTract} object.}

\item{pointType}{A string, either \code{"control"} (the default) or \code{"knot"},
indicating whether the spline's control points or its fitted knot
locations should be returned.}
}
\value{
A list with elements \code{points}, a matrix of 3D point coordinates,
and \code{seedPoint}, the index of the seed point within \code{points}.
}
\description{
This function extracts either the control points or the knot locations of
a \link{BSplineTract}, along with the index of its seed point, in the common
form used by \code{\link[=calculateSplineStepVectors]{calculateSplineStepVectors()}} and related functions.
}
\author{
Jon Clayden
}
