% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25_tract_models.R
\name{readPntDataTable}
\alias{readPntDataTable}
\title{Read a PNT data table}
\usage{
readPntDataTable(datasetName)
}
\arguments{
\item{datasetName}{A string giving the name of the dataset, with or
without a ".txt" extension.}
}
\value{
A data frame containing the dataset.
}
\description{
This function reads a probabilistic neighbourhood tractography (PNT)
dataset, as created by \code{\link[=createDataTableForSplines]{createDataTableForSplines()}}, from file. If the
complete dataset file does not exist, but pieces of it created separately
(for example by \code{plough}) do, these are combined, the combined file is
written out, and the pieces are removed.
}
\author{
Jon Clayden
}
