% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_nt.R
\name{transformStreamlineWithOptions}
\alias{transformStreamlineWithOptions}
\title{Transform a streamline into reference space}
\usage{
transformStreamlineWithOptions(options, streamline, session,
  refSession = NULL)
}
\arguments{
\item{options}{A \code{tractOptions} object, as created by
\code{\link[=createTractOptionList]{createTractOptionList()}}.}

\item{streamline}{A \link[tractor.track:Streamline]{tractor.track::Streamline} object to transform.}

\item{session}{The \link[tractor.session:MriSession]{tractor.session::MriSession} that \code{streamline} was
generated from.}

\item{refSession}{An optional \link[tractor.session:MriSession]{tractor.session::MriSession} object giving
the session to register \code{streamline} to. The default, \code{NULL}, registers
to standard (MNI) space instead.}
}
\value{
The transformed streamline, returned invisibly. If
\code{options$registerToReference} is \code{FALSE}, \code{streamline} is returned
unchanged.
}
\description{
This function transforms a streamline into the space of a reference
session, or into standard (MNI) space, if requested by the given tract
generation options.
}
\author{
Jon Clayden
}
