% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{writeNTResource}
\alias{writeNTResource}
\title{Write a PNT resource to file}
\usage{
writeNTResource(object, type, options = NULL)
}
\arguments{
\item{object}{The object to serialise: usually a \link{ReferenceTract},
\link{UninformativeTractModel}, \link{MatchingTractModel} or
\link{ProbabilisticNTResults} object.}

\item{type}{A string, one of \code{"reference"}, \code{"model"} or \code{"results"},
giving the type of resource being written.}

\item{options}{A named list of elements used to construct the resource's
file name; see \code{\link[=getFileNameForNTResource]{getFileNameForNTResource()}}.}
}
\value{
The serialised representation of \code{object}, as returned by
\code{\link[tractor.base:serialiseReferenceObject]{tractor.base::serialiseReferenceObject()}}, invisibly. This function is
mainly called for its side effect of writing the file.
}
\description{
This function serialises a probabilistic neighbourhood tractography (PNT)
reference tract, matching model or results object to file, at the location
determined by \code{\link[=getFileNameForNTResource]{getFileNameForNTResource()}}.
}
\author{
Jon Clayden
}
